+++ src/openepd/bundle/__init__.py
@@ -17,4 +17,3 @@
 #  Charles Pankow Foundation, Microsoft Sustainability Fund, Interface, MKA Foundation, and others.
 #  Find out more at www.BuildingTransparency.org
 #
-VERSION = "0.12.0"